requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
av==16.1.0
bidict==0.23.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
ctranslate2==4.7.1
curl_cffi==0.14.0
docutils==0.22.4
fastapi==0.125.0
faster-whisper==1.2.1
filelock==3.24.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.3.2
htmx==0.0.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ifaddr==0.2.0
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
mdurl==0.1.2
MouseInfo==0.1.3
mpmath==1.3.0
mss==10.1.0
multidict==6.7.1
nicegui==3.7.1
numpy==2.4.2
onnxruntime==1.24.2
orjson==3.11.7
outcome==1.3.0.post0
packaging==26.0
pillow==12.1.1
playwright==1.58.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
py-spy==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==1.10.26
pydantic_core==2.41.5
pyee==13.0.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pyparsing==3.3.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.2.1
python-engineio==4.13.1
python-multipart==0.0.22
python-socketio==5.16.1
python3-xlib==0.15
pytweening==1.2.0
PyWavelets==1.9.0
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
rsa==4.9.1
scipy==1.17.1
selenium==4.41.0
setuptools==82.0.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
typer==0.24.1
typing_extensions==4.15.0
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
yarl==1.22.0
yt-dlp==2026.3.3
